requirements.txt
numpy
matplotlib.pyplot
shapely
torch
loguru
json
gzip
pickle